requirements_video.txt
opencv-python
pillow
websockets
numpy